OpenAI-Agents-SDK/01-ai-agents-first/04-hello_agent/async/hello_agent/pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nida1khurram", email = "147607279+nida1khurram@users.noreply.github.com" }
]
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.0",
]

[project.scripts]
hello-agent = "hello_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
